refactor io modules into structured packages
