+++ jarvis/__init__.py
@@ -29,18 +29,10 @@ from .constants import (
     ExecutionResult,
 )
 
-try:
-    from .voice import (
-        PicovoiceWakeWordListener,
-        OpenAITTSEngine,
-        ElevenLabsTTSEngine,
-        VoiceInputSystem,
-    )
-except Exception:  # pragma: no cover - optional deps
-    PicovoiceWakeWordListener = None
-    OpenAITTSEngine = None
-    ElevenLabsTTSEngine = None
-    VoiceInputSystem = None
+PicovoiceWakeWordListener = None
+OpenAITTSEngine = None
+ElevenLabsTTSEngine = None
+VoiceInputSystem = None
 __all__ = [
     "CalendarService",
     "AIClientFactory",
@@ -61,10 +53,6 @@ __all__ = [
     "create_from_file",
     "DEFAULT_PORT",
     "LOG_DB_PATH",
-    "PicovoiceWakeWordListener",
-    "OpenAITTSEngine",
-    "ElevenLabsTTSEngine",
-    "VoiceInputSystem",
     "PROTOCOL_RESPONSES",
     "ExecutionResult",
 ]
