+++ jarvis/io/input/__init__.py
@@ -1,0 +1,13 @@
+"""Input handling components."""
+
+from .system import VoiceInputSystem
+from .wakeword.base import WakeWordListener
+from .transcription.base import SpeechToTextEngine
+from .wakeword.mock import MockWakeWordListener
+
+__all__ = [
+    "VoiceInputSystem",
+    "WakeWordListener",
+    "SpeechToTextEngine",
+    "MockWakeWordListener",
+]
