+++ jarvis/io/input/transcription/base.py
@@ -1,4 +1,3 @@
-# speech_to_text.py
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
