+++ jarvis/io/input/wakeword/__init__.py
@@ -1,0 +1,9 @@
+from .base import WakeWordListener
+from .mock import MockWakeWordListener
+
+try:  # optional dependency
+    from .picovoice import PicovoiceWakeWordListener
+except Exception:  # pragma: no cover - optional deps may be missing
+    PicovoiceWakeWordListener = None
+
+__all__ = ["WakeWordListener", "MockWakeWordListener", "PicovoiceWakeWordListener"]
