+++ jarvis/io/input/wakeword/base.py
