+++ jarvis/io/input/wakeword/picovoice.py
@@ -2,14 +2,14 @@ from __future__ import annotations
 
 import asyncio
 import os
-from typing import Iterable, Sequence
+from typing import Sequence
 
 import numpy as np
 import sounddevice as sd
 import pvporcupine
 
-from ..logger import JarvisLogger
-from .wake_word import WakeWordListener
+from ....logger import JarvisLogger
+from .base import WakeWordListener
 
 
 class PicovoiceWakeWordListener(WakeWordListener):
@@ -30,13 +30,9 @@ class PicovoiceWakeWordListener(WakeWordListener):
         self.logger = logger or JarvisLogger()
         self.debug = debug
 
-        # Initialize Porcupine with error handling
         try:
             if not self.keyword_paths:
-                # Use built-in "jarvis" keyword if no custom keyword paths are provided
-                self._porcupine = pvporcupine.create(
-                    access_key=self.access_key, keywords=["jarvis"]
-                )
+                self._porcupine = pvporcupine.create(access_key=self.access_key, keywords=["jarvis"])
             else:
                 self._porcupine = pvporcupine.create(
                     access_key=self.access_key, keyword_paths=self.keyword_paths
@@ -45,23 +41,17 @@ class PicovoiceWakeWordListener(WakeWordListener):
             self.logger.log("ERROR", "Failed to initialize Porcupine", str(e))
             raise
 
-        # Don't initialize stream in constructor - do it when needed
         self._stream = None
 
     def _create_stream(self):
-        """Create audio input stream with error handling."""
+        """Create an audio input stream."""
         try:
-            # Check if audio devices are available
             devices = sd.query_devices()
             if not devices:
                 raise RuntimeError("No audio devices found")
-
-            # Find default input device
             default_input = sd.default.device[0]
             if default_input is None:
                 raise RuntimeError("No default input device found")
-
-            # Create the stream
             return sd.InputStream(
                 samplerate=self._porcupine.sample_rate,
                 channels=1,
@@ -73,12 +63,10 @@ class PicovoiceWakeWordListener(WakeWordListener):
             raise
 
     async def wait_for_wake_word(self) -> None:
-        """Wait until a wake word is detected."""
+        """Block until the wake word is detected."""
 
         def _detect() -> None:
-            # Create stream fresh each time to avoid "Invalid stream pointer" errors
             stream = self._create_stream()
-
             try:
                 with stream:
                     while True:
@@ -86,22 +74,15 @@ class PicovoiceWakeWordListener(WakeWordListener):
                             pcm, overflowed = stream.read(self._porcupine.frame_length)
                             if overflowed:
                                 self.logger.log("WARNING", "Audio buffer overflowed")
-
                             pcm = np.frombuffer(pcm, dtype=np.int16)
                             result = self._porcupine.process(pcm)
-
                             if result >= 0:
                                 if self.debug:
                                     self.logger.log("DEBUG", "Wake word detected")
                                 break
-
                         except Exception as e:
-                            self.logger.log(
-                                "ERROR", "Error during wake word detection", str(e)
-                            )
-                            # Break on any error during detection
+                            self.logger.log("ERROR", "Error during wake word detection", str(e))
                             break
-
             except Exception as e:
                 self.logger.log("ERROR", "Stream context error", str(e))
                 raise
@@ -110,13 +91,11 @@ class PicovoiceWakeWordListener(WakeWordListener):
             await asyncio.to_thread(_detect)
         except Exception as exc:
             self.logger.log("ERROR", "Wake word listener error", str(exc))
-            # Re-raise to let the caller handle it
             raise
 
     def __del__(self):
-        """Clean up resources."""
         if hasattr(self, "_porcupine") and self._porcupine:
             try:
                 self._porcupine.delete()
-            except:
+            except Exception:
                 pass
