+++ jarvis/io/output/__init__.py
@@ -1,0 +1,6 @@
+"""Output handling components."""
+
+from .tts.base import TextToSpeechEngine
+from .tts.mock import MockTTSEngine
+
+__all__ = ["TextToSpeechEngine", "MockTTSEngine"]
