+++ jarvis/io/output/tts/__init__.py
@@ -1,0 +1,16 @@
+from .base import TextToSpeechEngine
+from .mock import MockTTSEngine
+
+try:  # optional heavy deps
+    from .openai import OpenAITTSEngine
+    from .elevenlabs import ElevenLabsTTSEngine
+except Exception:  # pragma: no cover - optional deps may be missing
+    OpenAITTSEngine = None
+    ElevenLabsTTSEngine = None
+
+__all__ = [
+    "TextToSpeechEngine",
+    "MockTTSEngine",
+    "OpenAITTSEngine",
+    "ElevenLabsTTSEngine",
+]
